requirements.txt
requests 
python-nmap 
ipinfo
python-owasp-zap-v2.4
reportlab
whoisdomain
dnspython